backend/pyproject.toml
[project]
name = "music-discovery-backend"
version = "0.1.0"
description = "High-performance backend for Music Discovery Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "polars>=0.20.5",
    "psycopg[binary,pool]>=3.1.18",
    "pgvector>=0.2.4",
    "python-dotenv>=1.0.1",
    "kaggle>=1.6.3",
    "numpy>=1.26.3",
    "scikit-learn>=1.4.0",
    "httpx>=0.26.0",
    "sqlalchemy>=2.0.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "tqdm>=4.66.0",
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "scripts"]
